Migration to wavelink 3.1
