+++ .env.dist
@@ -1,0 +1,11 @@
+BOT_TOKEN = ""
+
+# Server access
+SERVER_IP=""
+SERVER_PORT=""
+SERVER_ENDPOINT=""
+
+# Wavelink
+WAVELINK_URL = ""
+WAVELINK_PORT = ""
+WAVELINK_PASSWORD = ""
