+++ requirements.txt
@@ -1,23 +1,23 @@
-aiohttp==3.8.4
+aiohttp==3.9.1
 aiosignal==1.3.1
-async-timeout==4.0.2
-attrs==22.2.0
-Brotli==1.0.9
-certifi==2022.12.7
-cffi==1.15.1
-charset-normalizer==3.1.0
+async-timeout==4.0.3
+attrs==23.1.0
+Brotli==1.1.0
+certifi==2023.11.17
+cffi==1.16.0
+charset-normalizer==3.3.2
 discord==2.3.2
 discord.py==2.3.2
-frozenlist==1.3.3
-idna==3.4
+frozenlist==1.4.1
+idna==3.6
 multidict==6.0.4
-mutagen==1.46.0
+mutagen==1.47.0
 pycparser==2.21
-pycryptodomex==3.17
+pycryptodomex==3.19.0
 PyNaCl==1.5.0
-websockets==10.4
-yarl==1.8.2
+websockets==12.0
+yarl==1.9.2
 python-dotenv==1.0.0
 static-ffmpeg==2.5
-wavelink==2.6.3
-sentry_sdk==1.17.0
+wavelink==3.1.0
+sentry-sdk==1.39.1
